requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
openai
pydantic
streamlit
python-dotenv